refactor(signer): simplify version transformer
keep protobuf the same as v1
