+++ packages/cita-sdk/examples/transfer.js
@@ -1,12 +1,12 @@
 const chalk = require('chalk')
-const web3 = require('./web3')
+const cita = require('./cita')
 const { privateKey } = require('./config')
 
 const divider = () => console.log(chalk.green('-'.repeat(10)))
 
-const account = web3.base.accounts.privateKeyToAccount(privateKey)
+const account = cita.base.accounts.privateKeyToAccount(privateKey)
 
-web3.base.getBalance(account.address).then(console.log)
+cita.base.getBalance(account.address).then(console.log)
 
 const transaction = {
   from: '0XB4061FA8E18654A7D51FEF3866D45BB1DC688717',
@@ -14,14 +14,14 @@ const transaction = {
   nonce: 999999,
   quota: 1e8,
   chainId: 1,
-  version: 1,
+  version: 2,
   validUntilBlock: 999999,
   value: ''
 }
 
 const transfer = async (to, value) => {
   checkBalance(to)
-  const current = await web3.base.getBlockNumber()
+  const current = await cita.base.getBlockNumber()
   const tx = {
     ...transaction,
     to,
@@ -30,7 +30,7 @@ const transfer = async (to, value) => {
   }
 
   console.log(chalk.green.bold(`Transaction to ${to} with value ${value}`))
-  const result = await web3.base.sendTransaction(tx)
+  const result = await cita.base.sendTransaction(tx)
   console.log(chalk.green.bold('Received Result:'))
   console.log(chalk.blue(JSON.stringify(result, null, 2)))
   setTimeout(() => {
@@ -39,7 +39,7 @@ const transfer = async (to, value) => {
 }
 
 const checkBalance = async to => {
-  const balance = await web3.base.getBalance(to, 'pending')
+  const balance = await cita.base.getBalance(to, 'pending')
   console.log(chalk.green.bold(`Now ${to} has balance of ${balance}`))
 }
