+++ packages/cita-signer/lib/index.js
@@ -40,19 +40,20 @@ const signer = ({ from, privateKey, data = '', nonce = exports.getNonce(), quota
     }
     let _to = to.toLowerCase().replace(/^0x/, '');
     let _chainId = chainId;
-    let _version = +version ? `V${version === 2 ? 1 : version}` : '';
+    let _version = +version ? `V1` : '';
     let _nonce = `${nonce}`;
     let _quota = +quota;
     switch (_version) {
-        case 'V1': {
+        case '': {
+            break;
+        }
+        case 'V1':
+        default: {
             _to = new Uint8Array(exports.hex2bytes(_to));
             _chainId = exports.hex2bytes('' + chainId);
             const chainIdBytes = new Uint8Array(32);
             chainIdBytes.set(_chainId, 32 - _chainId.length);
             _chainId = chainIdBytes;
-            break;
-        }
-        default: {
             break;
         }
     }
