+++ packages/cita-signer/lib/unsigner.js
@@ -20,13 +20,15 @@ const unsigner = (hexUnverifiedTransaction) => {
         validUntilBlock: +transactionPb.getValidUntilBlock(),
     };
     switch (+version) {
+        case 0: {
+            break;
+        }
         case 1:
-        case 2: {
+        case 2:
+        default: {
             transaction.chainId = '0x' + (+index_1.bytes2hex(transactionPb.getChainIdV1())).toString(16);
             transaction.to = index_1.bytes2hex(transactionPb.getToV1());
             break;
-        }
-        default: {
         }
     }
     if (transaction.to && !transaction.to.startsWith('0x')) {
